requirements.txt
wget
numpy
tensorflow_datasets
gin-config
tqdm
